pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentna"
version = "0.1.0"
description = "Local code agent with memory and change tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "AgentNA Team" }
]
keywords = ["ai", "code-agent", "memory", "rag", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # TUI
    "textual>=0.47.0",

    # Vector DB
    "chromadb>=0.4.22",

    # Knowledge Graph
    "networkx>=3.2",

    # LLM Providers
    "ollama>=0.3.0",
    "anthropic>=0.18.0",

    # Code Parsing
    "tree-sitter>=0.21.0",

    # File Watching
    "watchdog>=4.0.0",

    # Git Integration
    "GitPython>=3.1.40",

    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",

    # Data Models
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Configuration
    "pyyaml>=6.0.1",

    # Utilities
    "pathspec>=0.12.0",
    "httpx>=0.26.0",

    # MCP
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.scripts]
agent = "agentna.cli.main:app"
agentna = "agentna.cli.main:app"

[project.urls]
Homepage = "https://github.com/agentna/agentna"
Documentation = "https://github.com/agentna/agentna#readme"
Repository = "https://github.com/agentna/agentna"

[tool.hatch.build.targets.wheel]
packages = ["src/agentna"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
